Better terminal logging
When using `pipe-pane` you have to deal with ANSI escape characters. You
have to filter it and still it will look bad, because backspace and tab
completion will be also recorded so in the end it will look like a
mess. On the other hand `capture-pane` works exactly how I want, so you
don't have to filter anything. What you see on the screen is what you
will get in the `capture-pane` output.